pyproject.toml
[project]
name = "flow_production_tracking"
version = "0.1.0"
description = "Griptape Nodes for Flow Production Tracking"
authors = [
    {name = "Jason Schleifer", email = "jason@griptape.ai"}
]
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "griptape-nodes",
    "httpx",
    "Pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flow_production_tracking"]
